requirements.txt
streamlit==1.28.0
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
transformers==4.34.1
torch==2.2.1
newspaper3k==0.2.8
pandas==2.1.1